Fixes Power Clip and Power Mask LPE
Some operations are missing on undo/redo previously getting some crashes and wrong on the undo stack.
To fix it I remove the undo insensitive code, yes, now two undos for one UI operation...
And add document done to be sure all steps are added to the undo stack. (previosly we miss last one of the verb)