On X11 with document units diferent to px text BBOX background of lower elements render scaled
When changing from one text item to start another the text render background scaled to document units
This happends on start creation a new text element. Is very often and you need to zoom to fix.
I test for example with a red circle in document ant click and type, but also happends in text in a frame in SVG2 mode
To fix this I check the click is not in the begin (new element created) of layout. I remove 99 % of the issues
You need to do a supper fast click type lots of times to reproduce again