Add user preference for redraw priority
Defaults to "Responsive" (G_PRIORITY_HIGH_IDLE), but can be set to the
previous value by selecting "Conservative" (G_PRIORITY_DEFAULT_IDLE).
Changes are effective immediately.

Persisted values are 100..200 and mapped to the actual priority range at
runtime. (For now, this amounts to a 1:1 mapping.) Custom values are supported if the user hacks them into the preference file.