Fix the top to bottom text conversion case.
When concerting tspans in older files, the cordinates for
tspans in top to bottom text is reversed. Making "x" -> "y"
when called getAttribute. But not reversed when calling
setAttribute. This is likely some internal part of Layout-TNG.

So this is a quick handle fix for this case with a comment.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/1227