Tool was not reset when document was swapped out
This started off trying to fix a slow down problem in issue #2328
but inn order to fix this the fix for issue #2031 has be reverted.

The crash issue needed to be fixed more generally. It turns out
that the event_context wasn't reset when the document changed.
This is bad, because it means ALL tools, not just the connector
tool, have unintended behaviours when swapping documents.

For example using the pen tool, starting a line and closing the
window results in the line magically moving to the new document.

This patch keeps the selected tool, but resets any in-progress actions.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/2031
Fixes https://gitlab.com/inkscape/inkscape/-/issues/2328