Calculate a meaningful threshold for boolops
Previously we would create shapes with many many nodes when
converting arcs to beziers. Mc added code that allowed the
operations to respect the threshold but these were set to
fixed amounts which didn't scale with the size of the objects.

This calculates thresholds for boolean operators so that they
have a balance between many nodes, and accuracy.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/2392