Avoid use of 'px' to specify UI styles
Replace most uses of 'px' used to specify sizes in the UI CSS with 'pt'.
Sizes specified in pixels will, in some cases of non-standard scaling /
high-DPI setups, result in physical pixels rather than any sensible unit
and can lead to text that is illegibly tiny. Units specified in 'pt' do
not have this problem.

Some '1px' values have been retained under the assumption these are
meant to mean "hairline" and that it is therefore acceptable for them to
be specified in physical pixels.

Also, remove unneeded units from zero values for consistency.