Prevent rapidly clicking on two different tools from opening preference dialog.
With actions, the button states of all buttons tied to the same action are automatically updated.
This removes the need to use Gtk::RadioToolButton's to ensure only one tool is shown as selected.
We can use Gtk::ToogleToolButton's instead. The advantage of this is that one doesn't get an activate
signal for the old button when a new button is clicked like one does with Gtk::RadioToolButton's.
This makes determining if a user double clicked on a tool icon to open the preference dialog more
robust.