Move warning of svg extension to failure branch
Currently we warn when the svg is successfully openned as
an svg, but doesn't have the svg extension. But we don't
warn when the fileis not opened.

This logic is backwards. We shouldn't warn when successfull and
should warn when unsuccessfull

Fixes Part 1. of https://gitlab.com/inkscape/inkscape/-/issues/2798