Cap coarsener parameters at half the tile size
...so that the algo doesn't get pushed outside of its design range by
people experimenting too much, and end up doing stupid things :)

Specifically, setting the coarsener min_size and glue_size larger than
the tile size (they should be smaller) and turning on slow mode to see
what's going on causes massively overlapping tiles to be generated. Some
people might come away thinking that the algo must therefore be
operating inefficiently all the time, when in fact it's just due to a
terrible choice of parameters. So we pre-emptively avoid that confusion.