Fix crash from the way outline mode was handled
Crash was due to outline store not being initialised, because the
criterion used by on_idle to determine whether it needed to be
initialised was different from the one on_draw used to draw it.

We also take this opportunity to clean up the way that this information
was communicated (previously via a round trip from Document).