Remove the ActionContext class.
This class served to track (in a complicated way) which Inkscape::Selection should be used for Verbs.
Gio::Actions (not related to the deleted class) use the "current selection" tracked by
InkscapeApplication. The current selection is either a selection owned by the Desktop of the
currently active InkscapeWindow or is a selection owned by the current document when there is no GUI.

This commit also removes several signals from Inkscape::Application that were redundant with other
signals and appeared not be actually used.