Correct transform handling by Boolean operations
The handling of affine transformations by the boolean operations on shapes
is corrected, so that these operations work correctly on paths which have
a clip or mask applied to them.

Context:
A boolean operation on paths requires that they first be converted to a
common coordinate system, and then the result must be converted back to
the coordinates of the "source" path (the bottom one in the Z-order).
This second step was done very crudely before, by simply setting the
`transform` attribute of the result node to an appropriate string,
overwriting the previous value.

Changes:
After these changes, the `transform` attribute is always copied from the
source object, and the remaining affine transformation is applied directly
to the result object. Preserving the `transform` attribute (which could
have been manually entered by the user and should not be altered without
user's consent) also preseves the relationship of the result object with
its clipPaths and masks, whenever the latter have the attribute
clipPathUnits="userSpaceOnUse" defined (which is the default).

Impact:
The boolean operations now work correctly on clipped and masked objects.
Since the Eraser Tool uses boolean operations internally, its "Cut" mode
can now correctly erase from clipped and masked shapes.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/1364
Fixes https://gitlab.com/inkscape/inkscape/-/issues/3236