Fix build with poppler 22.03.0
PDFDoc's constructor now takes an std::unique_ptr as the filename and
the rest of the parameters are std::optional and can be left out.

Also, remove an obsolete comment regarding special handling on win32;
the "ifndef _WIN32" it was referring to has been removed by commit
ad8effaa6ec3 ("Fix PDF import with non-ASCII filename on Windows").