UPDATED: added dX and dY measurements to the measure tool
I have only changed the files measure-tool.ccp and measure-tool.h
in src/ui/tools.

I managed to compile and test it on my laptop on Ubuntu 21.10.
I haven't been able to compile it on Windows yet (even the original/unchanged
version would not compile...) as there is an error recognizing the mingw64 configuration.

From the user point of view, the change is only in 2 extra measurements (dX and dY)
visible next to the total length measurement.