Record pen path in caligraphy for hatching
When using the hatching mode in the caligraphy pen, we might also want
to use the tracing functionality. But the thickness of the lines makes
the outline of the original path accumulate more problems the more
lines with a variable line width are added together.

To get around this, we record the pen's center path and use that for
hatching if it's available.

The new attribute isn't inteded to be used in any sp object.