Change SVG double formatting precision meaning
Fixes https://gitlab.com/inkscape/inkscape/-/issues/3947

The precision `tprec` parameter of `sp_svg_number_write_de()` means
differently for input number greater than 1 or smaller than 1. I suppose
the precision should mean the number of significant digits, but for
input smaller than 1, the `tprec` specifies the number of digits
following the decimal point. For example, `0.007619` formats as `0.0076`
given `tprec = 4`. This leads to unexpected precision loss for small
numbers.

Rewrite the function so that `tprec` refers to the number of significant
digits in all cases. Use double-conversion lib for formatting the number
instead of `std::stringstream` because there are rules like plus sign
before positive exponents and two exponent digits (1.6e+01) that I could
not find a way to control with the stdlib.