New UX to split out delete node tool actions
Instead of fighting modes, we split out the different delete node
modes into their own tool actions which a user can reconfigure to shortcuts
as they need to.

The preference which used to decide is removed.

The default delete accessed through toolbar buttons and menu items is has
a new preference drop down which can select any of the available modes.