Add popover menus to toolbar SpinButtons, restoring 1.3.x behavior (mostly).
These menus were previously used when SpinButtons fell off the end of toolbars.
We no longer use such menus. But they were also accessable by right-clicking on
the SpinButton.

For most SpinButtons, the popup menus are not that interesting but they are
better than the default popup menu which includes the ability to add emojis, etc.
Some menus are useful, such as with the Star tool where the "Spoke ratio" menu
has values that allows one to create pentagrams, hexagrams, etc.

The merger request also fixes the sharing of the same popover menu between
all spinbuttons (the variable referencing the menu was declared static in
the SpinButton class) which caused only the first opened popover menu to work.