Uncomment erasing comments while reading style (CSS Dialog)
The CSS dialog currently breaks style in files that have comments in the
css style node. In the original code from [1] had a block of code that
removed comments in CSS styles, which has hence been commented out.

Uncomment code to add it back, since it appears to have been done for
testing purposes.

Fixes: https://gitlab.com/inkscape/inkscape/-/issues/5218

[1] https://gitlab.com/inkscape/inkscape/-/merge_requests/619